Ensure all dbs are closed before deleting file
This was supposed to be fixed long ago, but SQLiteDataFile was doing
things in the wrong order: calling the super close method, which
decrements the file’s in-use count, before actually closing the SQLite
db. That meant that another thread calling deleteDatabase would
unblock and delete the file before the first thread was done closing.

Fixes #739