Changed query observer API.
The enumerator is now accessed by calling the new function
c4queryobs_getEnumerator(), which can be called at your leisure any
time after the notification. This lets you easily schedule the work
on your preferred thread, instead of trying to use the enumerator
on the random background thread that the observer is called on.