Fixed “datatype mismatch” error in ANY/EVERY query
SQLiteFleeceEach was failing to parse Fleece data at an odd address.
This was already worked around in the fix for #589 but that only
covered regular queries, not the special ‘virtual table’ used by ANY
and EVERY queries. This time I made sure that there are no more places
where this bug could lurk.

Fixes #787