Fixed C++ exception pushing huge numbers of revs
Pusher::revHistoryString() would sometimes write every revision to
the history, ignoring the maxRevTreeDepth limit. If there were
thousands of revs in the history, this could overflow the BLIP
message's properties and cause an exception on the receiving end.

I fixed this. (Also moved the method to RevToSend, since it makes
more sense there.)

On the receiving side, I changed RevTree::insertHistory so it won't
complain about truncated histories, as long as the truncated part is
revs that would be pruned anyway.

Unit tests added for both loopback and SG replications.

Fixes CBL-106