Fix for super-fast pulls with deltas (CBL-136)
If the puller receives a `changes` message with a docID that's already
got a pending rev coming in, don't return any ancestor rev IDs for it.
That way we won't receive a delta with a base revision that's no
longer current and we don't have the body of.
Fixes CBL-136