Fix for super-fast pulls with deltas (CBL-136)
(This is the Cobalt (2.6) version of the fix: the replicator has been
refactored since 2.5 so the original patch f3fb70e2 isn't applicable.)

If the puller receives a `changes` message with a docID that's already
got a pending rev coming in, don't return any ancestor rev IDs for it.
That way we won't receive a delta with a base revision that's no
longer current and we don't have the body of.

Fixes CBL-136

# Conflicts:
#	Replicator/ReplicatorTypes.cc