CBL-221 Keep Puller in busy status longer (#817)
It needs to wait until all the unfinished incoming revisions are done.  These are interjected in a thread-suspicious way by the Inserter.

Added test to ensure that document ended callbacks fire correctly.
There is no easy way to detect the underlying exception since it is caught and handled internally in the actor, so check the symptoms instead (some of the doc ended callbacks don't fire).  This test has a 100% failure rate on master, and 100% pass rate on this branch.