Fixed race condition starting replicators (#829)
It's possible for a replicator to start and call a progress callback
before the c4repl_new function returns; in that case the client
hasn't yet set the C4Replicator handle, so if the callback references
it, it will get null (or worse, garbage.)
This is a pretty rare situation given the timing required, but it
does happen to the ReplicatorAPITest occasionally on Jenkins and
causes a test failure.
To address this, I added a flag to the options that if set prevents
c4repl_new from starting the replicator. Then the client, once it has
the C4Replicator handle, can call a new function c4repl_start.