Better memory management of Fleece values from C4Documents
If you get an FLValue from a C4Document, and then either retain it
(via `FLValue_Retain()`) or add it to a mutable Array or Dict,
the FLValue will now remain valid until released, even if the owning
C4Document is released.
Most of the work was done in Fleece code, but I did have to change
VersionedDocument to keep Docs, not Scopes, since Scopes aren't
ref-counted.