Refactored checkpoint and pending-docs code
Moved SequenceSet into Checkpoint class

The Checkpoint class itself now keeps track of exactly which sequences
are pending.
There's also some experimental support for writing the set of
pending sequences into the checkpoint, but it's not ready yet.

More refactoring/cleanup of Checkpointer

Reimplemented SequenceSet with run-length encoding.

Correctly identify passive replicator objects in logs

Replicator objects are supposed to have their class names lowercased
in logs when they are passive, e.g. "pusher" instead of "Pusher.
Unfortunately the test for this (in Worker::loggingClassName) was
broken and only checked whether the pull side of the replication was
passive.
Fixed this by adding a `_passive` member to Worker. Subclasses are
responsible for setting it in their constructors.

Improved checkpoint representation again

No more minSequence/maxSequence, just the SequenceSet.

Track checkpoint as set of _completed_ sequences

Inverts the meaning of the SequenceSet: it's now the set of completed
sequences, not the set of pending sequences. This simplifies the code
since it doesn't have to deal with the weird final range that ends at
'infinity'.

Pusher: don't send sequences the checkpoint says are complete

This makes a difference when starting a replication from a saved
checkpoint containing discontiguous ranges. The pusher will start
from the first pending sequence, but there will be completed sequences
past there that shouldn't be sent again.
This also includes some optimizations to avoid allocating a RevToSend
if the revision's going to be skipped anyway.

Some fixes to Checkpointer

* Don't reload checkpoint from db every time read() is called
* Store Checkpoint as a pointer so we know if it's loaded and so
  we can remove the include of Checkpoint.hh
* C4 pendingDocIDs API should return a null slice when there are none,
  not an empty Fleece array.