Correctly identify passive replicator objects in logs
Replicator objects are supposed to have their class names lowercased
in logs when they are passive, e.g. "pusher" instead of "Pusher.
Unfortunately the test for this (in Worker::loggingClassName) was
broken and only checked whether the pull side of the replication was
passive.
Fixed this by adding a `_passive` member to Worker. Subclasses are
responsible for setting it in their constructors.