Pusher: don't send sequences the checkpoint says are complete
This makes a difference when starting a replication from a saved
checkpoint containing discontiguous ranges. The pusher will start
from the first pending sequence, but there will be completed sequences
past there that shouldn't be sent again.
This also includes some optimizations to avoid allocating a RevToSend
if the revision's going to be skipped anyway.