API: Added c4queryobs_setEnabled, to prevent a race condition
After a thread creates an observer, it was possible for the callback
to be called (on another thread) before c4queryobs_create returned.
In this case the app's reference to the observer wouldn't be set yet,
which could cause problems (as in one of our unit tests.)
To prevent this case, a query observer now has to be enabled before
its callback will be invoked.
This API also makes it possible to disable an observer without
freeing it, which could be useful in situations like a backgrounded
app.