Fix broken build from last commit (but break .NET build)
Restored the lost `C4API` and `static inline` declarations to the
retain/release functions in c4Base.h.

Removed the inline retain/release functions from the bindings list.

I know this is going to break the C# build because it needs generated
bindings for those retain/release functions; but this is necessary to
fix the other builds and unblock other developers, before Jim comes in
and can figure out how to make .NET work too.