Repl should copy C4SocketFactory in case ptr is invalidated
Fixed an invalid-memory issue reported by ASan in the iOS unit tests:
CBL is passing a C4ReplicatorParameters containing a pointer to a
temporary stack-based C4SocketFactory struct. The c4RemoteReplicator
hangs onto that pointer and tries to use it again later when retrying.

Fixed by having c4RemoteReplicator copy the C4SocketFactory struct
into a member variable if the ptr is non-null.