Fixed: Passing bad C4Database* to C4PredictiveModel
Incorrect conversion of DataFileDelegate* to c4Database* (due to the
intricacies of C++ multiple inheritance.)
Fixed by including the
header that defines the c4Database class, so the compiler knows how
much to offset the pointer.
Also made the cast into a `dynamic_cast` so it will produce a
compile error if it doesn't know the proper offset.

This was manifesting as a bogus C4BlobStorage* passed to
c4blob_getContents.

Fixes CBL-648