LiveQuerier: Fixed race condition
LiveQuerier was listening to the original Database's SequenceTracker,
so it got notified the moment a change happened, before it was
committed. But it queries on a background database, which won't see
any of the changes before the commit. So there was a race condition,
where if the LiveQuerier ran its query soon enough, it wouldn't see
any changes.

Fixed this by adding a transaction observer interface to BackgroundDB
so it can notify LiveQueries when transactions finish.

Also improved logging a bit.

Fixes CBL-688