Speculative fix for missing pulled blobs
Fixes an edge condition where, if the socket disconnects unexpectedly,
the replicator might add a pulled document even though its blobs
haven't been downloaded.
Fix is for IncomingBlob to set an error status when it fails to read
the entirety of its streamed BLIP message. That tells its parent
IncomingRev that everything is not OK, so it won't try to save the
revision.
We don't have any good way to unit-test this, unfortunately.

Fixes CBL-608