CBL-680: Prevent issues when deleting the logging directory
If the logging directory is deleted while the logger is active, the previous call to mkdir inside of the purge function would fail.  First of all, there is no reason to do a mkdir here so remove that, but more importantly the break in flow will cause dangling pointers which will bring the whole process down the next time a log is attempted.  Prevent that by nulling after deletion.