Fixed a crashing Timer bug
It was possible for a Timer instance to be rescheduled even while its
destructor was being called. The Timer would then be invoked later on
when it was garbage, likely causing a crash or corrupting whatever
other object occupied the addresses of the Timer's _triggered and
_state fields.

Sequence to reproduce:
1. Timer's callback is called (on Timer thread)
2. Timer's destructor is called (on another thread), blocks...
3. The callback calls timer->setFireTime() to reschedule
4. Callback exits
4. ...destructor exits now that callback is done.
5. Later, the Timer::Manager triggers the deleted Timer; boom.

Fixed this by having the Timer's destructor move it to a new kDeleted
state, and having setFireTime() do nothing if the timer is in that
state.

I believe this was the cause of CBL-854 (Sporadic SIGILL exceptions).
I was able to catch this problem with the Thread Sanitizer while
running c4DocumentTests, after messing with the timing of closing the
database vs. running expirations such that I got the expiration task
to run while the database was closing.

Fixes CBL-854