CBL-848: Don't allow two listeners on the same port (#954)
This has the side effect of not being able to restart a listener for up to 60 seconds on some platforms as their sockets linger in TIME_WAIT or whatever it is called.  

Also map the WSA codes to POSIX which is slightly complicated because strerror on Windows does not return valid messages for these.  The messages come from their WSA error counterparts.