Fix withPublicKey() when the public key is not in the KeyChain
The public key may not be in the KeyChain as a separate entity, for example, when using SecPKCS12Import to import the identity from P12 created from openSSL tool on macOS, the public key entity will not be presented in the KeyChain. When that happens, create the SecKeyRef of the public key from the public key data instead.

CBL-947