Fix cert validation on iOS (CBL-995)
iOS for some reason doesn't make the list of trusted root certs
available to 3rd party apps. This means we can't give mbedTLS the
list of roots, so it can't validate peer certificates.

However, mbedTLS _does_ have an API to register a callback that will
be asked to find the root cert that signed a given peer cert, and
we _can_ do that on iOS using SecTrust.

Fixes CBL-995