Disabled timeout of C4Test::importJSONLines
It had a default timeout of 15 sec, and was timing out when reading
the big iTunes library file with the Thread Sanitizer enabled.
That meant it was just returning early without reading all the docs,
which made the tests fail because the wrong number of docs were
replicated.

I changed it to never timeout unless an explicit timeout is given.