Improved threading in replicator tests
Changed all the sleep-loops in ReplicatorAPITest to waits on a
condition lock.

Changed timeouts to fail the test, since if the test continues after
a timeout it will start misbehaving in ways that aren't obviously
related to the timeout.

In ReplicatorLoopbackTest I made it grab the mutex _before_ starting
the replicators, so there aren't any race conditions there.