Fixed race condition in Puller
Puller::updateLastSequence calls replicator(), which fails an assertion
if the parent link has already been cleared. It was clearly trying to
call the replicator only if it's still available; I fixed it to do
that, by adding replicatorIfAny() which is like replicator() without
the assertion.

This led me to notice that going up the `_parent` chain isn't thread-
safe, since your parent could clear its parent link at any moment on
its actor thread. So I'm now doing it in a safer way that grabs a
retained reference at each step, and ends up returning a retained
reference to the Replicator.

Fixes CBL-1043