Fixed potential deadlock in pending-docs API
Blake discovered that c4repl_isDocumentPending calls the replicator's
validation callback while holding the replicator mutex, which can lead
to a deadlock if the callback calls something else that locks a mutex.

Fixed this by changing the implementation to lock the replicator,
grab the necessary state, unlock, then use the state to get the
results. It actually adds a little helper class encapsulating the
state, which seemed the cleanest way to do it.