CBL-1054: Don't do offline handling after c4repl_stop
This manifests in a very hard to hit situation where a replicator cannot connect to the other side because of a transient error, but since it has already switched to offline it has no connection to close and so the call to stop has no effect other than to change the status to stopping.  This needs to be detected in the status changed callback and the offline handling should be skipped.