CBL-1044 Add self-signed validation mode
This needs to live in harmony with the other modes, and furthermore it needs to disrupt normal processing by ignoring the normal CA validation.  Only self signed certificates will be accepted.  The motivation for this is due to our commitment to make an insecure mode as difficult to achieve as possible.  If this were simply a mode to allow self signed certificates in addition to valid ones, then this could easily be overlooked and put into production thereby making a product using the library easily attackable.  On the other side, many P2P implementations will have dynamically generated self signed certificates floating around and those should work by default if that mode is desired.  So the basic thinking is that in P2P the self signed mode will be used, and with non-P2P the normal mode will be used.  It is possible to use the self signed mode with non-P2P scenarios but this would be a poor idea in production.  It could be used during development, but there would be an actual error when switching to production if this mode was left on, thereby preventing the accidental insecurity.