CBL-1089: Pusher should ignore conflicted documents
There are two steps involved in this.  Ignoring conflicted documents is easy enough but then there is a potential that they are ignored later because the resolution process does not create any new sequences (i.e. the remote revision is accepted as-is).  So the documentSaved method needs to be "delayed" by both ignoring it for conflicts, and then calling it when certain conditions are met in the resolve conflict method.