CBL-1061: Always set delegate to null in terminate() (#1026)
This issue is not reproducible, but static analysis leads me to believe that what is happening is that C4LocalReplicator is getting destroyed via c4repl_free() at a point when the replicator's connection is already gone.  This means that since connected() will be false, the delegate will remain non-null.  Furthermore, static analysis shows that this is always the object which is getting freed so there is no point in keeping it around after a call to terminate.