Fix a deadlock that was revealed in the LogEncoder
The LogEncoder destructor holds a mutex, and then deletes its flush timer.  If the timer has been triggered but not yet fired in this case a deadlock will occur because the destructor is waiting for the timer callback to complete, but the timer callback is waiting for the mutex that the destructor is holding.  It is safe to delete the timer before acquiring the mutex because usually all that happens is that the timer removes itself from scheduling.  The actual flush operation, however, still needs to be guarded since it uses shared objects.