Revert removing assertion and fix underlying cause of failure
Regression from CBL-1089 caused this to happen.  The revision does not need to be added to the sequence tracker at that moment, since resolveConflict is not meant to save anything.  It just needs to reset the sequence back to 0 so that it can get a new one on the follow up save() call.