CBL-936: Make sure Puller's counters react to a failed delta (#1047)
When a delta failure happens, the error response triggers the revision to be sent again as a full revision, but without going through the changes message first.  The changes message both adds progress and also adds pending rev message count that is balanced when the revision is received, and so needs to be re-added as part of the revision re-request.

* Modify LiteCore to behave the same as Sync Gateway

When a delta fails with a 422, Sync Gateway will send the full revision without another changes feed entry, but LiteCore sends it with another changes feed entry.  These two ways of doing things are incompatible because of the unbalanced progress and counter bumps that happen on the active side when a 422 occurs