CBL-1163: Add note about the format of the data in the sign callback
It's very confusing to otherwise try to reason about it because the data being received is opaque and there are two ways to handle it.  One is to hash and sign the data and the other is to simply sign the data (which is what needs to be done in this case).