API Change: Removed C4Revision.body, added c4doc_getBody and c4doc_getRoot
Starting a migration towards accessing the body as a FLDict rather
than a slice. This will allow a document to store multiple revisions
in one Fleece body.
Step one is to remove the `body` field from C4Revision, and instead
make it available from a function, `c4doc_getBody()`.
Additionally, added `c4doc_getRoot()` to get the revision root as a
FLDict directly.