This commit introduces the main changes
It starts using CMake with more of the power it comes with in order to inspect what threading functions are available, so that ThreadUtil.hh can properly detect the name and numeric ID of the current thread, if called.  This allows GetThreadName to function in a way that returns a thread name that is unique enough to distinguish (on Linux, for example, threads inherit the process name by default and so it is hard to distinguish between different default-named threads, so add the numeric ID too).

Armed with this, the new ChannelManifest class has the ability to "follow" a trail of actors as they pass messages between each other and show this information in the event of an exception as a sort of "actor stack trace."  This works in the same way for both GCDMailbox and ThreadedMailbox and the process is described in ChannelManifest.hh