Modify c4_setTempDir and eliminate FilePath::tempDirectory
This has caused issues on some platforms that do not have a reliable "default" temp directory.  It needs to be on the same filesystem of a destination that is only known at runtime.  c4_setTempDir now *only* sets the SQLite temp directory, and it is now an error to call it twice.

This commit breaks the tests, but in order to keep things wheaty and dead simple this commit is a record of things that are important to the library itself.  The next commit will address the tests.

Moving the temp directory logic up to the test level plowed me into some technical debt regarding `using namespace std` in headers which causes frustrating compile errors (especially on Windows).  There are a few fixups, including a Fleece bump because of an incompatibility I noticed that will be addressed in the commit after that.