Modify c4_setTempDir and eliminate FilePath::tempDirectory
Backport of #1086

This has caused issues on some platforms that do not have a reliable "default" temp directory.  It needs to be on the same filesystem of a destination that is only known at runtime.  c4_setTempDir now *only* sets the SQLite temp directory, and it is now an error to call it twice.

The main changes are around this, and the rest is the result of incompatibilities and technical debt that was exposed by this change (80% of which is shifting around `using namespace std` declarations)