Adapt the rest of the actors to the changes
This would have made the previous commit noisy, so this is part 2 of the last commit, which applies the changes in the previous commit to the rest of the actors and adds a "test" (doesn't actually test anything, but maybe that can be changed.  It's mostly to observe the behavior of the new system).