Improved non-null annotations in C API
Now using Clang's `_Nullable` annotation, along with the pragmas
`clang assume_nonnull begin` and `clang assume_nonnull end`. This is
the same mechanism used in Objective-C APIs. Its advantage over
`__attribute((nonnull))` is that it does not imply that passing null
is undefined behavior, so it doesn't affect code optimization in ways
that make it impossible to add an explicit null check inside the
function.

Because of this way this stuff works, the headers now explicitly tag
_nullable_ pointers, instead of ones that cannot be null.

Pointer fields in structs also cannot be NULL except when tagged as
nullable.

This stuff is wrapped in macros for compatibility: in between
C4_ASSUME_NONNULL_BEGIN and C4_ASSUME_NONNULL_END, all pointer
declarations *implicitly* disallow NULL values, unless annotated with
C4NULLABLE (which must come after the `*`.)

An explicit C4NONNULL is occasionally necessary when there are
multiple levels of pointers.