CBL-1515: Aggregate the platform variations of URL for Checkpoint
Cherry-picked backport of #1102

Some platforms were passing in an explicit port, even for default 80/443, while others were just leaving it blank. This led to discrepencies between the platform checkpoints, as a remote URL is used for calculating a replication checkpoint.  If a preloaded database is used on a platform that differs in this manner, then the previous checkpoint will fail to be detected.  This commit solves this by checking for both variations during the previous checkpoint search.