Optimized c4doc_update()
Restored the "fast path" that writes the new record to the db without
reading first. This works as long as the doc was loaded with all revs
and has not been updated in the db.
If the fast path fails, it goes back to the prior behavior of calling
`c4doc_put`.
As part of this, cleaned up `c4doc_put` somewhat.