Deprecate keyword, META, in N1QL, for designating the meta property in property path. #CBL-1632
We deprecate the special property paths for the meta properties, such as "meta.id", "meta.sequence", etc. in N1QL. Instead, we have a new function, meta(), that returns an object of the meta-keys and the corresponding values. To get a particular meta property, such as, "meta.id", one uses nested operator, "meta().id".
  The special syntax for meta key in JSON, such as "._id", remain valid. However, N1QL will no longer generate this style of referring to meta properties in Json.