Fixed uncaught-exception crash in c4db_openNamed etc.
c4db_openNamed, c4db_exists, c4db_deleteNamed would crash with an
uncaught C++ exception if the input database name or parentDirectory
were a NULL slice. That's because the dbPath() subroutine would fail
an assertion, and so throw an exception, but the call to dbPath()
was not inside a try/catch block. Fixed.