Fix MSVC build issue, and implicit dependency
There is a bug in the MSVC 15.x compiler that Microsoft will not fix that prevents the use of guaranteed copy elision with deleted move constructors in some cases.  Use a workaround until upgrade to 16.x.

Also don't force define the cert requesting stuff in C4Certificate since it creates an additional dependency on another library when building consumers of the library.