Fix C4Document vs C4Document_C struct layout
This is a horrible kludge that needs to make some fields line up
between a C struct and a C++ class. It worked on Mac/Clang and
Linux/GCC, but broke on Windows/MSVC because apparently it has subtly
different rules for field layout in classes with inherited members.
This is my attempt to fix it.