Fix flaky "Continuous Push From Both Sides" test (#1170)
The test expects there to be an error pushing its one document.
Most of the time there is, but if the two replicators interleave
their changes, the test can finish with no conflicts. For whatever
reason this has become more likely of late -- I like to think it's
because of my replicator and CRUD optimizations, but who knows.

This patch adds a new boolean flag to ReplicatorLoopbackTest that,
when set, makes it OK for there to be no document errors even if
the expected-errors sets are non-empty. This flag is set in the one
test that exhibits this problem.