Restore old semantics of c4doc_update (returns copy)
Having c4doc_update return the same C4Document instance was a bad
idea:
* If the save failed, the document was left with the new revision
  inserted.
* It didn't match the platform code's expectations that changes to the
  original doc affected the new one.
To restore the original semantics I had to make a copy of the
C4Document object, and add the new revision to that and save it.
I also marked update() as `const` to enforce that it can't change the
document state.