Skip the duplicated version vector tests. #CBL-2000
For all test cases in c4Test and ReplicatorLoopbackTest, each test case is duplicated for RevTree and VersionVector. Since VersionVector is not the release focus for now, we skip them for now. Following are the time differences on my Mac,
c++ tests: 303s vs 510s, and
c   tests: 106s vs 161s.