Fixed: 1st param of MATCH() doesn't need a db alias
E.g. "FROM":[{"AS":"testdb"}],"WHERE":["MATCH()", "sentence", "words"]
is legal ... a regression had made it require "testdb.sentence".

I improved QueryParser::FTSTableName to parse the FTS index name with
the same logic as for regular property accesses, so it uses the same
rules for when a DB alias is needed.