Reading doc extra shouldn't replace body
When a C4Document is loaded without its 'extra', and then a later
call causes 'extra' to be loaded, the KeyStore updates all the fields
of the Record including the 'body', swapping it out with a new
alloc_slice with the same data.
This means that after this call, the document's old body as used by
clients is no longer valid, causing nasty errors.
This commit changes Record's setters to ignore a change to 'body' or
'extra' that leaves the data the same.