Implemented PropertyEncryption [API]
API addition: Defined encryptable-property constants in
c4Document+Fleece.h and c4Document.hh.

The PropertyEncryption class takes a document's properties as a Dict
and applies property-level encryption or decryption.

Its implementation is in the couchbase-lite-core-EE repo; there is a
stub source file here that either #includes it or defines a stub CE
implementation that does nothing but return an error if an encryptable
property is found.