Fix using already released revID slice in TreeDocument’s purgeRevision(revID)
TreeDocument’s purgeRevision(revID) is called by the resolveConflict() function to purge the losing revision. The purgeRevision(revID) function calls updateFlags() which may release the given revID if the given revID is the same as the current _revID.

This commit prevents the use of the already released revID from updateFlags() by moving the code that checks for isSelectedRevID above the updateFlags() call.

CBL-2160