Reworked BuiltInWebSocket handling of EOF on read (CBL-2204)
Fix for cases where BuiltInWebSocket won't close the connection when the peer
closes the socket unexpectedly:
- Peer calls shutdown
- BuiltInWebSocket::readFromSocket is called, reads 0 bytes, calls
  inherited close()
- WebSocketImpl::close sends a CLOSE message and waits for a reply.
- Five seconds later it times out. closeSocket() calls socket->close(), which
  invokes the `shutdown` system call and Poller::interrupt.
- Poller::interrupt tries to call the listeners on that FD, but there aren't
  any, so BuiltInWebSocket never finds out the socket is closed.

I added a new kDisconnect event for Poller, so the BIWS can be notified
reliably when the socket closes.

Also, it doesn't make sense to send a CLOSE message in this situation (peer
shutdown) since the peer can't possibly reply to it. Instead, WebSocketImpl
should  treat it as a protocol violation to receive an EOF before the peer's sent
a CLOSE. I added an `onEOF` method to signal this.