Fix unexpected purging after push/pull tomestones, CBL-2345 (#1245) (#1249)
Fix unexpected purging after push/pull tomestones, CBL-2345
As a deleted doc is pushed up, we lose the channel info because the deleted docs have empty bodies. This causes SGW to think all channels are removed from this doc and hence send down the Removal flag when pulled. This commit disregards the Removal flag if it is accompanied by the Deleted flag.