CBL-2364: Prevent a rare crash situation in external transaction notifier
The simplified flow of this is that while a DatabaseImpl is being destroyed, another Database operating on the same file could commit a transaction at just the right time and it would find its way into calling a method on the collection of the database being destroyed.  The collection on the database would be invalid at this point and cause a crash because `_keyStore` is null.