Tagging the opened database connections. CBL-2357 (#1253)
Tagging the opened database connections. CBL-2357
By default, a database connection is tagged "appOpened" as it is opened. These database connections are opened by the application, and, as
such, the application must close all of them as it calls to delete the underlying database. Other tags are reserved for the connections opened
by the Core library. The library is responsible to close them.
    Also changed kOtherDBCloseTimeoutSecs from 3 seconds to 6 seconds for CBL-2383.