Support HTTP chunked-mode responses, fixing timeout error (CBL-2313) (#1238)
* Support HTTP chunked-mode responses

The unit test "HTTPS Request to public host" has begin failing, not because of
a change in our code but because the couchbase.com server has started using
HTTP chunked mode to send the response. TCPStream doesn't recognize this
response type, so (because there's no Content-Length) it tries to read all data
till EOF. Meanwhile the server is using keep-alive, so after sending the
response it waits for another request on the socket. Deadlock! Which ends with
a timeout error.

Fixed by making TCPSocket recognize "Transfer-Encoding: chunked" and decoding
the output. While I was at it, I made it fail on other transfer encodings, and
made it only read till EOF if there is an explicit "Connection: close" header.

Note this is unlikely to affect CBL because the only HTTP requests it sends are
WebSocket connections, which don't have any regular response.

Fixes CBL-2313

Add RESTClientTest.cc to CMake project of CppTests

Co-authored-by: Jianmin Zhao <jianmin.zhao@couchbase.com>