CBL-2485: Account for a race between pendingDocs and terminate
There is a gap in between the pending document class checking for a valid replicator, and actually using the replicator later, in which the replicator could potentially become invalid and cause a null dereference.  The method where the issue was found has been modified to be able to distinguish between a good state and bad state via return values.  pendingDocumentIDs will return false if it couldn't run, and isDocumentPending will return nullopt.