CBL-2486: Verify couchbasedeps submodule commits (#1274)
From now on until further notice all submodules will need to conform to the same level of branch strictness as our main repos.  That means the following:

* Any repos in the couchbasedeps org (which usually do not change often) will either need to be on a tagged commit, or on a branch that either begins with "couchbase-" or ends with "-couchbase" (e.g. `couchbase-master` or `2.0.0-couchbase` is ok, `fix/couchbase-cool-stuff` is not)
* Any other repos must be on a branch with the same name as the target of the current PR.

This is to combat the frequent occurrence of making two PRs at once (one in a submodule and one in a parent), then merging them both without updating the parent to the new post-merge commit of the submodule.  This has the issue that it can easily end up with the parent pointing at an "orphan" commit in the submodule if the branch is deleted (as it should be).  Not only will this break any further builds but it leaves open the possibility that one day the orphaned commit is garbage collected and the product at that commit becomes unbuildable in the future.