Cbl 2477: Tear down DBAccess on stopped instead of on release (#1287)
CBL-2477: Tear down DBAccess on stopped instead of on release
This task is to port the fix in hotfix/2.8.7, filed as 2407, into Lithium. The code base has undergone non-trivial changes since Hydrogen(2.8.7); sheer cherry-picking can hardly do it. The idea is to release reference to the database as soon as the replicator reaches Stopped state, as supposed when all replicator's workers are released.