CBL-2543: Don't open DataFile in BackgroundDB too early
As it stands, the call to dataFile()->openAnother will register the datafile as a candidate to receive notifications of transactions committed by other handles but at the point that it does this the object is not fully constructed.  If a notification came in before the construction is finished this causes undefined behavior like hangs and crashing.