CBL-2562: Try to put down some long running race conditions (#1299)
* CBL-2562: Try to put down some long running race conditions

The replicator will close its database after it is stopped, but RevFinder continues trying to use it.  Wire up RevFinder so that it blocks the Replicator from stopping until RevFinder is finished.

* Catch and warn about situation where RevFinder can't find Replicator

Instead of hitting an assertion failure, it should log that it was unable to call back to the Replicator.  This is probably because the replicator is already stopped, so it is not likely to matter (The actual value was saved in the database already).

* Make sure revfinder status is initialized to idle

Otherwise in the case where the replication ends so fast that the rev finder never changes status, the replication will hang because it will never become idle.

Co-authored-by: Jim Borden <jim.borden@couchbase.com>